Refactor codebase and improve user authentication process
This commit includes changes across the entire codebase aimed to enhance the user experience and streamline operations. The main feature of this update is the introduction of user authentication via token insertion which is then used for all subsequent API requests. User input is now required if the authentication token is blank. Continuous enhancement activities include the replacement of $.ajax calls with axios library in SubscriptionPlans.vue and PaymentPage.vue and the addition of comprehensive details about each subscription plan in Home.vue. Furthermore, the subscription status is colour-coded for an improved user interface.